requirements.txt
django==1.8.2
MYSQL-python==1.2.5
django-tastypie
django-cors-headers
requests
whoosh
